Rename StoryListPage to StoryList
